Fix widget not updating on manual NAV update
For some reason I was not allowing it to update the widget if the NAV
udpate was manually launched. I said the updateWidget function would be
called from somewhere in the main thread but I wasn't able to find it. So
I removed the conditional update. Widget now updates always from the
service.